unify development utilities
