--- web/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-	<title>Example website for a Locutus-compatible web</title>
-</head>
-
-<body>
-	<p>This is an example website. <a href="./state.html">This</a> is a subpage for this website.</p>
-</body>
-
-</html>
